Fix for fractional container width rounding
element.clientWidth rounds the width value (see
https://developer.mozilla.org/en-US/docs/Web/API/Element/clientWidth).
If the width is rounded up, then rowGrid.js overestimates how much
space it has and the last image can wrap to its own row.

This fix uses getBoundingClientRect() to get the true width of the
container. The width property of the DOMRect object wasn’t included
until IE9, so for compatibility with IE8 this fix computers the width
using the right and left properties.